updated the scripts. still needs to be tested

+++ scripts/update_slide.sh
@@ -1,10 +1,11 @@
 #!/bin/bash
 
+cd ..
+
 git checkout gh-pages
 wget https://github.com/GDGBremen/slides-statusreport-$1/archive/$1.zip
-git rm slides-statusreport-$1
+if [ -d ./slides-statusreport-$1];
+then
+   git rm slides-statusreport-$1
+fi
 unzip slides-statusreport-$1.zip && rm slides-statusreport-$1.zip
-git add -u
-git add .
-git commit -m "updated gh-pages through update script"
-git push
